refactor: dao 추상화
