+++ app/src/main/java/com/techcourse/dao/UserDaoJdbc.java
@@ -1,0 +1,55 @@
+package com.techcourse.dao;
+
+import com.techcourse.domain.User;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+
+import javax.sql.DataSource;
+import java.util.List;
+import java.util.Optional;
+
+public class UserDaoJdbc implements UserDao {
+
+    private static final RowMapper<User> USER_ROW_MAPPER = rs -> new User(rs.getLong(1), rs.getString(2), rs.getString(3), rs.getString(4));
+
+    private final JdbcTemplate jdbcTemplate;
+
+    public UserDaoJdbc(DataSource dataSource) {
+        this.jdbcTemplate = new JdbcTemplate(dataSource);
+    }
+
+    public UserDaoJdbc(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    @Override
+    public void insert(User user) {
+        var sql = "insert into users (account, password, email) values (?, ?, ?)";
+        jdbcTemplate.update(sql, user.getAccount(), user.getPassword(), user.getEmail());
+    }
+
+    @Override
+    public void update(User user) {
+        var sql = "update users set (account, password, email) = (?, ?, ?) where id = ?";
+        jdbcTemplate.update(sql, user.getAccount(), user.getPassword(), user.getEmail(), user.getId());
+    }
+
+    @Override
+    public List<User> findAll() {
+        var sql = "select id, account, password, email from users";
+        return jdbcTemplate.query(sql, USER_ROW_MAPPER);
+    }
+
+    @Override
+    public Optional<User> findById(Long id) {
+        var sql = "select id, account, password, email from users where id = ?";
+        return jdbcTemplate.queryForObject(sql, USER_ROW_MAPPER, id);
+    }
+
+    @Override
+    public Optional<User> findByAccount(String account) {
+        var sql = "select id, account, password, email from users where account = ?";
+        return jdbcTemplate.queryForObject(sql, USER_ROW_MAPPER, account);
+
+    }
+}
