+++ app/src/main/java/com/techcourse/dao/UserHistoryDao.java
@@ -1,35 +1,8 @@
 package com.techcourse.dao;
 
 import com.techcourse.domain.UserHistory;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.jdbc.core.JdbcTemplate;
 
-import javax.sql.DataSource;
+public interface UserHistoryDao {
 
-public class UserHistoryDao {
-
-    private static final Logger log = LoggerFactory.getLogger(UserHistoryDao.class);
-
-    private final JdbcTemplate jdbcTemplate;
-
-    public UserHistoryDao(DataSource dataSource) {
-        this.jdbcTemplate = new JdbcTemplate(dataSource);
-    }
-
-    public UserHistoryDao(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    public void log(final UserHistory userHistory) {
-        final var sql = "insert into user_history (user_id, account, password, email, created_at, created_by) values (?, ?, ?, ?, ?, ?)";
-        jdbcTemplate.update(sql,
-                userHistory.getUserId(),
-                userHistory.getAccount(),
-                userHistory.getPassword(),
-                userHistory.getEmail(),
-                userHistory.getCreatedAt(),
-                userHistory.getCreateBy()
-        );
-    }
+    void log(final UserHistory userHistory);
 }
