+++ app/src/main/java/com/techcourse/dao/UserHistoryDaoJdbc.java
@@ -1,0 +1,30 @@
+package com.techcourse.dao;
+
+import com.techcourse.domain.UserHistory;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.jdbc.core.JdbcTemplate;
+
+public class UserHistoryDaoJdbc implements UserHistoryDao {
+
+    private static final Logger log = LoggerFactory.getLogger(UserHistoryDaoJdbc.class);
+
+    private final JdbcTemplate jdbcTemplate;
+
+    public UserHistoryDaoJdbc(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    @Override
+    public void log(final UserHistory userHistory) {
+        final var sql = "insert into user_history (user_id, account, password, email, created_at, created_by) values (?, ?, ?, ?, ?, ?)";
+        jdbcTemplate.update(sql,
+                userHistory.getUserId(),
+                userHistory.getAccount(),
+                userHistory.getPassword(),
+                userHistory.getEmail(),
+                userHistory.getCreatedAt(),
+                userHistory.getCreateBy()
+        );
+    }
+}
