+++ app/src/test/java/com/techcourse/service/MockUserHistoryDao.java
@@ -1,11 +1,11 @@
 package com.techcourse.service;
 
-import com.techcourse.dao.UserHistoryDao;
+import com.techcourse.dao.UserHistoryDaoJdbc;
 import com.techcourse.domain.UserHistory;
 import org.springframework.dao.DataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
 
-public class MockUserHistoryDao extends UserHistoryDao {
+public class MockUserHistoryDao extends UserHistoryDaoJdbc {
 
     public MockUserHistoryDao(final JdbcTemplate jdbcTemplate) {
         super(jdbcTemplate);
